// src/types/users.ts
export type UserProp = {
  uuid: string;
  id: string;
  password: string;
  createdAt: string;
  updatedAt: string;
  nickName: string;
  lastLoginTime: string;
  switchLevel: string;
  image: string;
  walletId: string;
  walletImg: string;
  concesHist: string;
  transAcc: string;
  kakaoUUID: string;
};